+++ core/src/com/unciv/ui/screens/worldscreen/unit/actions/UnitActionsPillage.kt
@@ -31,7 +31,9 @@ object UnitActionsPillage {
                 "Pillage",
                 true
             ) {
-                (pillageAction.action)()
+                GUI.getWorldScreen().runAndRecordSimultaneousGameStateChange(UnitActionType.Pillage) {
+                    (pillageAction.action)()
+                }
                 GUI.setUpdateWorldOnNextRender()
             }.open()
         })